+++ docs/conf.py
@@ -30,7 +30,7 @@ copyright = '2022, Konstantinos Lampridis'
 author = 'Konstantinos Lampridis'
 
 # The full version, including alpha/beta/rc tags
-release = '1.7.5'
+release = '1.8.0'
 
 # -- General configuration ---------------------------------------------------