+++ src/cookiecutter_python/__init__.py
@@ -1,1 +1,3 @@
-__version__ = '1.7.5'
+__version__ = '1.8.0'
+
+from . import _logging  # noqa