+++ src/cookiecutter_python/backend/__init__.py
@@ -1,3 +1,4 @@
+from .gen_docs_common import get_docs_gen_internal_config
 from .main import generate
 from .post_main import CheckWebServerError
 from .sanitization import sanitize
@@ -6,4 +7,5 @@ __all__ = [
     'generate',
     'CheckWebServerError',
     'sanitize',
+    'get_docs_gen_internal_config',
 ]