+++ tests/test_generate.py
@@ -3,6 +3,17 @@ import typing as t
 import pytest
 
 
+@pytest.fixture
+def assert_scaffolded_without_cli(module_file) -> t.Callable[[str], None]:
+    from os import path
+
+    def assert_project_generated_without_cli(project_dir: str) -> None:
+        get_file: t.Callable[[str], str] = module_file(project_dir)
+        assert all(not path.isfile(get_file(file_name)) for file_name in CLI_RELATED_FILES)
+
+    return assert_project_generated_without_cli
+
+
 @pytest.mark.parametrize(
     'config_file, expected_interpreters',
     [
@@ -50,6 +61,15 @@ def test_supported_python_interpreters(
 
 @pytest.fixture
 def assert_interpreters_array_in_build_matrix() -> t.Callable[[str, t.Sequence[str]], None]:
+    """Test that Job Matrix is generated correctly and stored as Workflow env var.
+
+    Test proper generation of github workflow config yaml for lines such as:
+
+    # FULL_MATRIX_STRATEGY: "{\"platform\": [\"ubuntu-latest\", \"macos-latest\", \"windows-latest\"], \"python-version\": [\"3.7\", \"3.8\", \"3.9\", \"3.10\", \"3.11\"]}"
+
+    Returns:
+        t.Callable[[str, t.Sequence[str]], None]: [description]
+    """
     from pathlib import Path
 
     def _assert_interpreters_array_in_build_matrix(
@@ -58,8 +78,9 @@ def assert_interpreters_array_in_build_matrix() -> t.Callable[[str, t.Sequence[s
     ) -> None:
         p = Path(project_dir) / '.github' / 'workflows' / 'test.yaml'
         contents = p.read_text()
-        b = ', '.join((f'"{int_ver}"' for int_ver in interpreters))
-        assert f"python-version: [{b}]" in contents
+        b = ', '.join((fr'\"{int_ver}\"' for int_ver in interpreters))
+        assert r'\"python-version\": ' in contents
+        assert fr'\"python-version\": [{b}]' in contents, f'"{b}" not in "{contents}"'
 
     return _assert_interpreters_array_in_build_matrix
 
@@ -68,6 +89,7 @@ CLI_RELATED_FILES = {
     'cli.py',
     '__main__.py',
 }
+"Files, only expected to be generated for cli type of Projects"
 
 
 @pytest.fixture
@@ -94,17 +116,6 @@ def module_file():
     return build_get_file_path
 
 
-@pytest.fixture
-def assert_scaffolded_without_cli(module_file) -> t.Callable[[str], None]:
-    from os import path
-
-    def assert_project_generated_without_cli(project_dir: str) -> None:
-        get_file: t.Callable[[str], str] = module_file(project_dir)
-        assert all(not path.isfile(get_file(file_name)) for file_name in CLI_RELATED_FILES)
-
-    return assert_project_generated_without_cli
-
-
 @pytest.fixture(params=[x for x in CLI_RELATED_FILES])
 def cli_related_file_name(request):
     return request.param
@@ -115,7 +126,9 @@ def test_enabling_add_cli_templated_variable(
     module_file,
     project_dir,
 ):
+    """Test that 'module+cli' Project Type generates CLI-explicit files."""
     from os import path
 
     get_file = module_file(project_dir)
+    assert path.exists(get_file(cli_related_file_name))
     assert path.isfile(get_file(cli_related_file_name))